\name{cgFraction}
\alias{cgFraction}
\title{ 
Percentage of C and G bases in a DNA sequence 
}
\description{
Computes the percentage of C and G bases in a DNA sequence 
}
\usage{
cgFraction(dna)
}
\arguments{
  \item{dna}{ 
a character string representing a DNA sequence 
}
}
\value{
  A numeric value in 0..100, giving the percentage of C and G bases.
}
