\name{expressionMatrixToDElist}
\alias{expressionMatrixToDElist}
\title{
Turn a Matrix of Gene Expression into a List of Differential Expression Tables
}
\description{
Transforms a matrix of gene expression values into a list of simple differential
gene expression data frames.  This is a reformatting utility for functions that 
operate on DE lists.
}
\usage{
expressionMatrixToDElist(x, groups = colnames(x), minimumValue = NULL, AVG.FUN = mean)
}
\arguments{
  \item{x}{
a numeric matrix of gene expression values, with rownames as the GeneIDs and column
names for the SampleIDs
}
  \item{groups}{
a character vector of group identifiers of the same length as the column width of \code{x}
}
  \item{minimumValue}{
a small linear offset to add to all expression values, before taking the log2 of the 
ratios, to prevent divide by zero and compensate for very low expression values
}
  \item{AVG.FUN}{
the math function for averaging gene expression per group
}
}
\details{
Many functions operate on DE lists, as generated by various differential expression tools
that operate on transcriptome file sets.  This utility function generates the same
format, but starting from a matrix of gene expression values.
}
\value{
A list of DE data frames, with a named element for each unique group in \code{groups}.
}
\author{
Bob Morrison
}
\seealso{
  \code{\link{geneSetAnalysis}} for a typical example of a function that uses DE lists
}
