\name{getVargeneDomainMap}
\alias{getVargeneDomainMap}
\alias{getVSAdomainMap}
\alias{getVSAgeneMap}
\title{
Variable Surface Antigen Maps
}
\description{
Specialized Plasmodium falciparum feature maps for variable surface antigens.
}
\usage{
getVargeneDomainMap()
getVSAdomainMap()
getVSAgeneMap()
}
\details{
These functions return detailed tables of specific vargenes for analysis
of P.falciparum datasets.
}
